[IMP][REF] reserve_without_save_sale,sale_product_customize: Correcciones y refactorizaciones varias

+++ project-addons/reserve_without_save_sale/controllers/main.py
@@ -21,8 +21,6 @@
 
 import json
 import logging
-import werkzeug
-import werkzeug.utils
 from openerp.addons.web import http
 from openerp.addons.web.http import request
 
@@ -31,26 +29,35 @@ _logger = logging.getLogger(__name__)
 
 class WebsiteReservation(http.Controller):
 
-
     @http.route(['/reservations/create'],
                 type='http', methods=['POST'], auth='user')
     def create_reservation(self, **post):
         _logger.debug('Creando reserva: %s', post)
         cr, uid, context = request.cr, request.uid, request.context
         reserv_obj = request.registry['stock.reservation']
-        warehouse  = request.registry['stock.warehouse'].browse(cr, uid, int(post['warehouse']), context)
-        vals = {
-            'product_id': int(post['product_id']),
-            'product_uom': int(post['uom']),
-            'product_uom_qty': float(post['qty']),
-            'date_validity': False,
-            'name': post['name'],
-            'location_id': warehouse.lot_stock_id.id,
-            'price_unit': float(post['price_unit']),
-            'unique_js_id': post['unique_js_id']
-        }
-        new_reservation = reserv_obj.create(cr, uid, vals, context)
-        reserv_obj.browse(cr, uid, new_reservation, context).reserve()
+        product_obj = request.registry['product.product']
+        sale_line_obj = request.registry['sale.order.line']
+        warehouse = request.registry['stock.warehouse'].\
+            browse(cr, uid, int(post['warehouse']), context)
+        product = product_obj.browse(cr, uid, int(post['product_id']), context)
+        if product.type == "product":
+            vals = {
+                'product_id': int(post['product_id']),
+                'product_uom': int(post['uom']),
+                'product_uom_qty': float(post['qty']),
+                'date_validity': False,
+                'name': post['name'],
+                'location_id': warehouse.lot_stock_id.id,
+                'price_unit': float(post['price_unit']),
+                'unique_js_id': post['unique_js_id']
+            }
+            new_reservation = reserv_obj.create(cr, uid, vals, context)
+            reserv_obj.browse(cr, uid, new_reservation, context).reserve()
+            line_ids = sale_line_obj.search(cr, uid, [('unique_js_id', '=',
+                                                       post['unique_js_id'])])
+            if line_ids:
+                reserv_obj.write(cr, uid, new_reservation,
+                                 {'sale_line_id': line_ids[0].id})
         return json.dumps(True)
 
     @http.route(['/reservations/unlink'],
@@ -59,7 +66,9 @@ class WebsiteReservation(http.Controller):
         unique_js_id = int(post['unique_js_id'])
         cr, uid, context = request.cr, request.uid, request.context
         reserv_obj = request.registry['stock.reservation']
-        reservation = reserv_obj.search(cr, uid, [('unique_js_id', '=', unique_js_id)], context=context)
+        reservation = reserv_obj.search(cr, uid, [('unique_js_id', '=',
+                                                   unique_js_id)],
+                                        context=context)
         reserv_obj.unlink(cr, uid, reservation, context)
         return json.dumps(True)
 
@@ -67,7 +76,8 @@ class WebsiteReservation(http.Controller):
                 type='http', methods=['POST'], auth='user')
     def write_reservation(self, **post):
         """
-            Si se ha modificado el producto, se elimina la reserva y se crea una nueva.
+            Si se ha modificado el producto, se elimina la reserva y se crea
+            una nueva.
             Si solo se modifica la cantidad se escribe.
 
             Si se modifica el producto:
@@ -82,23 +92,40 @@ class WebsiteReservation(http.Controller):
         """
         cr, uid, context = request.cr, request.uid, request.context
         reserv_obj = request.registry['stock.reservation']
+        product_obj = request.registry['product.product']
         if post.get('product_id', False):
-            reservation = reserv_obj.search(cr, uid, [('unique_js_id', '=', post['old_unique_js_id'])], context=context)
-            new_data = {
-            'name': post['name'],
-            'product_id': post['product_id'],
-            'unique_js_id': post['unique_js_id']
-            }
-            if post.get('qty'):
-                new_data['product_uom_qty'] = post['qty']
-            reserv_id = reserv_obj.copy(cr, uid, reservation[0], new_data, context)
-            reserv_obj.browse(cr, uid, reserv_id, context).reserve()
-            print('Copiada reserva')
-            reserv_obj.unlink(cr, uid, reservation, context)
-            print('eliminada reserva')
+            product = product_obj.browse(cr, uid, int(post['product_id']),
+                                         context)
+            if product.type == "product":
+                reservation = reserv_obj.\
+                    search(cr, uid, [('unique_js_id', '=',
+                                      post['old_unique_js_id'])],
+                           context=context)
+                new_data = {
+                    'name': post['name'],
+                    'product_id': post['product_id'],
+                    'unique_js_id': post['unique_js_id']
+                }
+                if post.get('qty'):
+                    new_data['product_uom_qty'] = post['qty']
+                reserv_id = reserv_obj.copy(cr, uid, reservation[0], new_data,
+                                            context)
+                reserv_obj.browse(cr, uid, reserv_id, context).reserve()
+                reserv_obj.unlink(cr, uid, reservation, context)
             return json.dumps(True)
-        if post.get('qty'):
-            reservation = reserv_obj.search(cr, uid, [('unique_js_id', '=', post['unique_js_id'])], context=context)
-            reserv_obj.write(cr, uid, reservation, {'product_uom_qty': post['qty']}, context)
-            print('modificada reserva')
+        elif post.get('qty'):
+            reservation = reserv_obj.search(cr, uid, [('unique_js_id', '=',
+                                                       post['unique_js_id'])],
+                                            context=context)
+            if reservation:
+                reserv_obj.write(cr, uid, reservation,
+                                 {'product_uom_qty': post['qty']}, context)
+        elif post.get('old_unique_js_id', False) and \
+                post.get('unique_js_id', False) and \
+                post['old_unique_js_id'] != post['unique_js_id']:
+            reservation = reserv_obj.\
+                    search(cr, uid, [('unique_js_id', '=',
+                                      post['old_unique_js_id'])],
+                           context=context)
+            reserv_obj.unlink(cr, uid, reservation, context)
         return json.dumps(True)
